+++ plugins/source/aws/resources/services/ssm/documents_mock_test.go
@@ -53,6 +53,23 @@ func buildSSMDocuments(t *testing.T, ctrl *gomock.Controller) client.Services {
 		},
 		nil,
 	)
+
+	var v types.DocumentVersionInfo
+	if err := faker.FakeObject(&v); err != nil {
+		t.Fatal(err)
+	}
+	v.Name = &docName
+	mock.EXPECT().ListDocumentVersions(
+		gomock.Any(),
+		&ssm.ListDocumentVersionsInput{Name: &docName},
+		gomock.Any(),
+	).Return(
+		&ssm.ListDocumentVersionsOutput{
+			DocumentVersions: []types.DocumentVersionInfo{v},
+		},
+		nil,
+	)
+
 	return client.Services{Ssm: mock}
 }
