+++ website/tables/aws/aws_ssm_documents.md
@@ -6,6 +6,11 @@ https://docs.aws.amazon.com/systems-manager/latest/APIReference/API_DocumentDesc
 
 The primary key for this table is **arn**.
 
+## Relations
+
+The following tables depend on aws_ssm_documents:
+  - [aws_ssm_document_versions](aws_ssm_document_versions)
+
 ## Columns
 
 | Name          | Type          |
